requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.1
email-validator==2.1.1
pytest==8.0.1
black==24.1.1
flake8==7.0.0